requirements.txt
absl-py==2.1.0
aiofiles==23.1.0
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2022.12.7
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.0
cycler==0.11.0
exceptiongroup==1.1.2
fastapi==0.101.0
ffmpy==0.3.1
filelock==3.12.2
flatbuffers==24.3.25
fonttools==4.41.0
frozenlist==1.4.0
fsspec==2023.6.0
gdown==5.2.0
google-auth==2.28.1
google-auth-oauthlib==1.0.0
gradio==3.39.0
gradio_client==0.3.0
grpcio==1.62.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
imageio==2.35.1
imgviz==1.7.5
importlib-metadata==7.0.1
importlib-resources==6.0.0
Jinja2==3.1.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
labelme==5.5.0
lazy_loader==0.4
linkify-it-py==2.0.2
Markdown==3.5.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
natsort==8.4.0
netron==7.5.2
networkx==3.1
numpy==1.24.2
oauthlib==3.2.2
onnx==1.13.1
onnxruntime==1.19.2
opencv-python==4.8.0.74
orjson==3.9.4
packaging==23.1
pandas==2.0.3
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.1.1
pydantic_core==2.4.0
pydub==0.25.1
pyparsing==3.0.9
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.6
python-opencv==1.0.0.14
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
QtPy==2.4.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.9.2
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
semantic-version==2.10.0
setuptools==69.1.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.6
starlette==0.27.0
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2023.7.10
toolz==0.12.0
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tornado==6.4
tqdm==4.65.1
typing_extensions==4.7.1
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.0.4
uvicorn==0.23.2
visdom==0.2.4
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.1
wheel==0.38.4
wincertstore==0.2
yarl==1.9.2
you-get==0.4.1650
zipp==3.16.2
